Extract src-tauri/src/usage.rs — everything read out of ~/.claude
The transcripts and the token ledger, one module: two consumers of one directory
that share the caveat CLAUDE.md puts on both — the layout is internal to Claude
Code and documented as unstable, so every reader here is a fallback chain rather
than a schema. list_past_sessions, read_transcript, transcript_meta,
project_transcript_dir, scan_usage, parse_usage_line, model_family and
token_usage_by_day, plus 6 tests. lib.rs 1,622 -> 1,097.

PLAN pairs this item with an injectable base dir so scan_usage,
list_past_sessions and read_transcript become testable against a fixture tree.
That changes signatures, so per ground rule 2 it is the next commit, not this
one; the module doc says so.

Two imports in lib.rs needed cfg gating rather than deleting: home_dir and (in
the test mod) scratch_dir. Each has exactly one user left and it is inside a
`cfg(target_os = "macos")` block — the legacy-localStorage reader and its test —
so an ungated import is an unused-import warning on Windows and Linux but not on
macOS. Found by the cfg flip added in the previous slice; the only warning it
still reports is the known reveal_path `exists` artifact.

Verification:
- usage.rs vs the pre-slice lib.rs, line-sorted: identical, 495 non-blank lines,
  no differences at all.
- the remainder of lib.rs: 12 lines — mod usage;, the 3 usage:: qualifications,
  and the two cfg gates with their comments.

cargo test: 69 passed, same as baseline. cargo check --tests clean, no warnings,
clean under the cfg flip.
